+++ spec/system/deployment_as_vendor_spec.rb
@@ -68,6 +68,10 @@ ceedling_system_tests do
       test_case :test_project_preprocessing_same_dir_mock_pairing
     end
 
+    describe "Partials" do
+      test_case :test_project_partial_all_module_mocks_declaration_only_function
+    end
+
     describe "Defines and configuration" do
       test_case :test_project_with_per_file_defines
       test_case :test_project_with_test_and_vendor_defines
@@ -181,6 +185,10 @@ ceedling_system_tests do
       test_case :test_project_with_preprocessing_for_missing_mock
       test_case :test_project_with_preprocessing_all
       test_case :test_project_preprocessing_same_dir_mock_pairing
+    end
+
+    describe "Partials" do
+      test_case :test_project_partial_all_module_mocks_declaration_only_function
     end
 
     describe "Defines and configuration" do